third_party/WebKit/LayoutTests/fast/events/scale-and-scroll-body-expected.txt

PASS successfullyParsed is true

TEST COMPLETE
FAIL window.document.scrollingElement.scrollTop should be 30. Was 0.
FAIL window.document.scrollingElement.scrollLeft should be 30. Was 0.
PASS window.document.scrollingElement.scrollTop is 100
PASS window.document.scrollingElement.scrollLeft is 100
PASS window.scrollX is 100
PASS window.scrollY is 100
